+++ clock/clock_test.go
@@ -1,0 +1,11 @@
+package clock
+
+import "testing"
+
+func TestClockFormat(t *testing.T) {
+	// TODO: DateFormat
+	// TODO: TimeFormat
+	// TODO: HourMinFormat
+	// TODO: StandardFormat
+	// TODO: TimestampFormat
+}