+++ clock/time_test.go
@@ -1,0 +1,7 @@
+package clock
+
+import "testing"
+
+func TestTime(t *testing.T) {
+	// TODO: DateTimeNow
+}