+++ CHANGELOG.md
@@ -2,6 +2,7 @@
   * Commit:
   * Request type
   * Request methods
+  * Clock format functions
 ## v0.3.0 - Rename Package
   * Commit: 2026-03-28 21:56
   * Rename package to github.com/zeroibot/pack